cloud_functions/requirements.txt
functions-framework
google-cloud-bigquery
google-cloud-firestore
google-cloud-logging
google-cloud-pubsub
google-cloud-language
